Introduce IEventSourcing and update version handling
The `IAggregate` interface in `IAggregate.cs` was modified to extend the new `IEventSourcing` interface. The `WithVersion` method in `EventDomain.cs` and `IEventDomain.cs` was updated to use `ulong` instead of `ushort` for the version parameter. A new `IEventSourcing` interface was introduced in `IEventSourcing.cs`, providing methods for handling events, including marking events as committed, getting uncommitted events, loading events from history, and pushing new events. The `Aggregate` class in `Aggregate.cs` was introduced, implementing the `IAggregate` interface and providing concrete implementations for the methods defined in `IEventSourcing`. This class includes methods for handling domain events, registering event handlers, and managing the state of the aggregate. Additionally, copyright and licensing information was added to `IEventSourcing.cs` and `Aggregate.cs`.

+++ Xpandables.Net/Events/EventDomain.cs
@@ -56,6 +56,6 @@ public abstract record EventDomain<TAggregate, TAggregateId> :
     public required TAggregateId AggregateId { get; init; }
 
     /// <inheritdoc/>
-    public virtual IEventDomain WithVersion(ushort version) =>
+    public virtual IEventDomain WithVersion(ulong version) =>
         this with { EventVersion = version };
 }
